+++ Example/Tests/TokenDSDK_Tests/APIv2/RequestBuilders/BaseJSONAPIRequestBuilderTests.swift
@@ -25,7 +25,7 @@ class BaseJSONAPIRequestBuilderTests: XCTestCase {
         )
         
         return JSONAPI.BaseApiRequestBuilderStack(
-            apiConfiguration: apiConfiguration,
+            apiConfigurationProvider: apiConfiguration,
             requestSigner: requestSigner,
             network: network
         )